Update MapReduce examples
