+++ batch/mapreduce-migration/src/main/java/WordCountCoreApi.java
@@ -14,32 +14,24 @@
  * limitations under the License.
  */
 
-import com.hazelcast.jet.AbstractProcessor;
+import com.hazelcast.jet.AggregateOperations;
 import com.hazelcast.jet.DAG;
 import com.hazelcast.jet.Jet;
 import com.hazelcast.jet.JetInstance;
-import com.hazelcast.jet.Traverser;
+import com.hazelcast.jet.Processors;
 import com.hazelcast.jet.Vertex;
 
-import javax.annotation.Nonnull;
-import java.util.HashMap;
-import java.util.Map;
 import java.util.Map.Entry;
-import java.util.StringTokenizer;
 
+import static com.hazelcast.jet.AggregateOperations.summingToLong;
 import static com.hazelcast.jet.Edge.between;
+import static com.hazelcast.jet.Partitioner.HASH_CODE;
+import static com.hazelcast.jet.Traversers.traverseArray;
 import static com.hazelcast.jet.function.DistributedFunctions.entryKey;
 import static com.hazelcast.jet.function.DistributedFunctions.wholeItem;
-import static com.hazelcast.jet.Partitioner.HASH_CODE;
-import static com.hazelcast.jet.Processors.readMap;
-import static com.hazelcast.jet.Processors.writeMap;
-import static com.hazelcast.jet.Traversers.lazy;
-import static com.hazelcast.jet.Traversers.traverseIterable;
 
 /**
- * Word count sample implemented without relying on out-of-the-box processors.
- * Demonstrates the lower-level convenience API to implement a cooperative
- * processor.
+ * Word count sample implemented with Hazelcast Jet's Core API.
  */
 public class WordCountCoreApi {
     public static void main(String[] args) throws Exception {
@@ -48,11 +40,15 @@ public class WordCountCoreApi {
         JetInstance jet = Jet.newJetInstance();
         try {
             DAG dag = new DAG();
-            Vertex source = dag.newVertex("source", readMap("sourceMap"));
-            Vertex map = dag.newVertex("map", MapP::new);
-            Vertex reduce = dag.newVertex("reduce", ReduceP::new);
-            Vertex combine = dag.newVertex("combine", CombineP::new);
-            Vertex sink = dag.newVertex("sink", writeMap("sinkMap"));
+            Vertex source = dag.newVertex("source", Processors.readMap("sourceMap"));
+            Vertex map = dag.newVertex("map", Processors.flatMap((String line) ->
+                    traverseArray(line.toLowerCase().split("\\W+"))
+                            .filter(word -> !word.isEmpty())));
+            Vertex reduce = dag.newVertex("reduce", Processors.groupByKey(
+                    wholeItem(), AggregateOperations.counting()));
+            Vertex combine = dag.newVertex("combine", Processors.groupByKey(
+                    entryKey(), summingToLong(Entry<String, Long>::getValue)));
+            Vertex sink = dag.newVertex("sink", Processors.writeMap("sinkMap"));
             dag.edge(between(source, map))
                .edge(between(map, reduce).partitioned(wholeItem(), HASH_CODE))
                .edge(between(reduce, combine).partitioned(entryKey()).distributed())
@@ -60,67 +56,6 @@ public class WordCountCoreApi {
             jet.newJob(dag).execute().get();
         } finally {
             Jet.shutdownAll();
-        }
-    }
-
-    private static class MapP extends AbstractProcessor {
-        private final FlatMapper<Entry<Long, String>, String> flatMapper = flatMapper(
-                (Entry<Long, String> e) -> new WordTraverser(e.getValue())
-        );
-
-        @Override
-        protected boolean tryProcess0(@Nonnull Object item) {
-            return flatMapper.tryProcess((Entry<Long, String>) item);
-        }
-    }
-
-    private static class WordTraverser implements Traverser<String> {
-
-        private final StringTokenizer tokenizer;
-
-        WordTraverser(String document) {
-            this.tokenizer = new StringTokenizer(document.toLowerCase());
-        }
-
-        @Override
-        public String next() {
-            return tokenizer.hasMoreTokens() ? tokenizer.nextToken() : null;
-        }
-    }
-
-    private static class ReduceP extends AbstractProcessor {
-        private final Map<String, Long> wordToCount = new HashMap<>();
-        private final Traverser<Entry<String, Long>> resultTraverser =
-                lazy(() -> traverseIterable(wordToCount.entrySet()));
-
-        @Override
-        protected boolean tryProcess0(@Nonnull Object item) {
-            wordToCount.compute((String) item, (x, count) -> 1 + (count != null ? count : 0L));
-            return true;
-        }
-
-        @Override
-        public boolean complete() {
-            return emitFromTraverser(resultTraverser);
-        }
-    }
-
-    private static class CombineP extends AbstractProcessor {
-        private final Map<String, Long> wordToCount = new HashMap<>();
-        private final Traverser<Entry<String, Long>> resultTraverser =
-                lazy(() -> traverseIterable(wordToCount.entrySet()));
-
-        @Override
-        protected boolean tryProcess0(@Nonnull Object item) {
-            final Entry<String, Long> e = (Entry<String, Long>) item;
-            wordToCount.compute(e.getKey(),
-                    (x, count) -> e.getValue() + (count != null ? count : 0L));
-            return true;
-        }
-
-        @Override
-        public boolean complete() {
-            return emitFromTraverser(resultTraverser);
         }
     }
 }
